refactor: test에 사용하는 객체의 hashtag수 증가

+++ hashtagmap-instagram-scheduler/src/main/java/com/songpapeople/hashtagmap/scheduler/CrawlingResult.java
@@ -14,7 +14,7 @@ import java.util.stream.Collectors;
 
 @NoArgsConstructor
 public class CrawlingResult {
-    public static final int MIN_HASHTAG_COUNT = 500;
+    public static final long MIN_HASHTAG_COUNT = 500;
 
     private CrawlingDto crawlingDto;
     private Place place;
